Enum exar::DatabaseError

pub enum DatabaseError {
    AuthenticationError,
    ConnectionError,
    EventStreamError(EventStreamError),
    IoError(ErrorKind, String),
    ParseError(ParseError),
    SubscriptionError,
    ValidationError(ValidationError),
}

A list specifying categories of database error.

Variants
AuthenticationError

The credentials used to connect to the database are either missing or invalid.

ConnectionError

The connection to the database failed.

EventStreamError(EventStreamError)

The event stream has been closed unexpectedly.

IoError(ErrorKind, String)

An I/O error occurred.

ParseError(ParseError)

The parsing of an event from the log file failed.

SubscriptionError

The attempted subscription failed.

ValidationError(ValidationError)

The validation of the event failed.

Methods

impl DatabaseError

fn from_io_error(err: IoError) -> DatabaseError

Returns a DatabaseError from the given std::io::Error.