Struct exar::Event

pub struct Event {
    pub id: u64,
    pub data: String,
    pub tags: Vec<String>,
    pub timestamp: u64,
}

Exar DB's event.

Examples

extern crate exar;

use exar::*;

let event = Event::new("data", vec!["tag1", "tag2"]);

Fields
id: u64

The event id (or sequence number).

data: String

The event data/payload.

tags: Vec<String>

The event tags.

timestamp: u64

The event timestamp.

Methods

impl Event

fn new(data: &str, tags: Vec<&str>) -> Event

Returns a new Event with the given data and tags.

fn with_id(self, id: u64) -> Self

Returns a modified version of the event by setting its id to the given value.

fn with_timestamp(self, timestamp: u64) -> Self

Returns a modified version of the event by setting its timestamp to the given value.

fn with_current_timestamp(self) -> Self

Returns a modified version of the event by setting its timestamp to the current time.